% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_assert.R
\name{c_assert}
\alias{c_assert}
\title{Check assertions about data}
\usage{
c_assert(data, predicate, source, ...)
}
\arguments{
\item{data}{A data frame.}

\item{predicate}{A function that returns FALSE when violated.}

\item{source}{Logical: TRUE or FALSE.}

\item{...}{Comma-separated list of unquoted columns.}
}
\value{
If the predicate assertion is true, TRUE.

If the predicate assertion is false and source is TRUE, a data-frame with the offending rows.

If the predicate assertion is false and source is FALSE, an assertr data-frame.
}
\description{
Check assertions about data.
}
