% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_tbl.R
\name{date_tbl}
\alias{date_tbl}
\title{Sub-set tables according to date range}
\usage{
date_tbl(x, col, from, to)
}
\arguments{
\item{x}{A multi-column table.}

\item{col}{A column of dates that are in year-month-day format.}

\item{from}{A start date in year-month-day format.}

\item{to}{An end date in year-month-day format.}
}
\description{
Sub-set multi-column tables so that one column of dates or date-times complies with a start date and an end date, inclusive.
}
\details{
In x, leading zeros in single-digit months or days, are optional.
}
\examples{
x <- data.frame(date_1 = c("1960-1-1", "1960-1-2", "1960-3-31"),
                date_2 = c("1960-1-1", "1960-1-2", "1960-3-31"))
z <- date_tbl(x = x, col = "date_1", from = "1960-1-1", to = "1960-3-30")
z
}
