% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_dir.R
\name{db_dir}
\alias{db_dir}
\title{Locations of embedded databases.}
\usage{
db_dir(folder, path)
}
\arguments{
\item{folder}{any folder above both 1) the database and 2) the R file.}

\item{path}{relative to \code{folder}, path to the database.}
}
\value{
A length-one character vector.
}
\description{
Locations of embedded databases.
}
