% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc}
\alias{dc}
\title{Date-time conversion to character}
\usage{
dc(x)
}
\arguments{
\item{x}{An object in year-month-day format, to be converted.}
}
\description{
Function to convert non-character representations of dates and date-times to objects of class "character" representing dates.
Executes strftime with assumptions about the values of some arguments.
}
\details{
In x, leading zeros in single-digit months or days, are optional.

The current time zone is used for the conversion.

The format method is "%Y-%m-%d".

Does not append the time zone abbreviation to the output.
}
\examples{
x <- c("1960-1-1", "1960-1-2", "1960-3-31", "1960-7-30")
z <- dc(x = x)
z
}
