% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt_d.R
\name{ddt_d}
\alias{ddt_d}
\title{Convert dates or date-times to Date-class dates}
\usage{
ddt_d(x)
}
\arguments{
\item{x}{An object in which values are in year-month-day format.
Leading zeros in single-digit months or days, are optional.}
}
\value{

}
\description{
Convert dates or date-times to Date-class dates.
}
\examples{
ddt_d(x = Sys.Date())
ddt_d(x = Sys.time())
}
\keyword{}
