% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt_ds.R
\name{ddt_ds}
\alias{ddt_ds}
\title{Convert dates or date-times to date strings}
\usage{
ddt_ds(x)
}
\arguments{
\item{x}{An object in year-month-day format.}
}
\value{

}
\description{
Execute strftime with assumptions about the values of some arguments.
}
\details{
In x, leading zeros in single-digit months or days, are optional.
format is "%Y-%m-%d".
tz is the current time zone.
usetz excludes time zone abbreviations from outputs.
}
\examples{
x <- c("1960-1-1", "1960-1-2", "1960-3-31", "1960-7-30")
z <- ddt_ds(x = x)
z
}
\keyword{}
