% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt_s.R
\name{ddt_s}
\alias{ddt_s}
\title{Convert dates or date-times to strings of dates.}
\usage{
ddt_s(x)
}
\arguments{
\item{x}{An object in which values are in year-month-day format.
Leading zeros in single-digit months or days, are optional.}
}
\value{

}
\description{
Convert dates or date-times to strings of dates.
}
\examples{
ddt_s(x = Sys.Date())
ddt_s(x = Sys.time())
}
\keyword{}
