% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt_tbl.R
\name{ddt_tbl}
\alias{ddt_tbl}
\title{Sub-set tables according to dates}
\usage{
ddt_tbl(tbl, col, from, to)
}
\arguments{
\item{tbl}{A table.}

\item{col}{A column.}

\item{from}{Start date in year-month-day format. Leading zeros in single-digit months or days, are optional.}

\item{to}{End date in year-month-day format. Leading zeros in single-digit months or days, are optional.}
}
\value{

}
\description{
Sub-set tables so that dates in one column are in a date range, inclusive.
}
\keyword{}
