% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dso_d.R
\name{dso_d}
\alias{dso_d}
\title{Convert days-since-origin to Date-class dates}
\usage{
dso_d(x, origin)
}
\arguments{
\item{x}{An object that stores days-since-origin.}

\item{origin}{An origin date in year-month-day format.}
}
\value{

}
\description{
Execute as.Date with assumptions about the values of some arguments.
}
\details{
format is un-necessary.
tz is blank.
}
\examples{
dso_d(x = 32768, origin = "1900-1-1")
}
\keyword{}
