% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtbl.R
\name{dtbl}
\alias{dtbl}
\title{Sub-set Tables According to Date Range}
\usage{
dtbl(tbl, col, from, to)
}
\arguments{
\item{tbl}{A multi-column table.}

\item{col}{A column of dates that are in year-month-day format.}

\item{from}{A start date in year-month-day format.}

\item{to}{An end date in year-month-day format.}
}
\description{
Sub-set multi-column tables so that a column of dates or date-times in the table complies with a start date and an end date, inclusive.
}
\details{
In tbl, leading zeros in single-digit months or days, are optional.
}
\examples{
x <- data.frame(date_1 = c("1960-1-1", "1960-1-2", "1960-3-31"),
                date_2 = c("1960-1-1", "1960-1-2", "1960-3-31"))
z <- dtbl(tbl = x, col = "date_1", from = "1960-1-1", to = "1960-3-30")
z
}
