% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{export_excel}
\alias{export_excel}
\title{Export data-frames as Microsoft Excel files}
\usage{
export_excel(x, dirname, rpath)
}
\arguments{
\item{x}{A data-frame, or named list of data-frames that will be sheets in the Excel file.}

\item{dirname}{A directory above 1) the Excel file and 2) the R file.}

\item{rpath}{Relative to \code{dirname}, path to the Excel file.}
}
\description{
Export data-frames as Microsoft Excel files.

To name sheets, set \code{x} to a named list of data-frames.
}
\seealso{
\code{\link{import_excel}}
}
