% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_feather.R
\name{export_feather}
\alias{export_feather}
\title{Write data-frames to Feather files.}
\usage{
export_feather(x, folder, path)
}
\arguments{
\item{x}{a data-frame to write to \code{path}.}

\item{folder}{any folder above both 1) the path to write to and 2) the R file.}

\item{path}{relative to \code{folder}, path to write to.}
}
\description{
Write data-frames to Feather files.
}
\seealso{
\code{\link{import_feather}}
}
