% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_rds.R
\name{export_rds}
\alias{export_rds}
\title{Write rds files}
\usage{
export_rds(object, dirname, rpath)
}
\arguments{
\item{object}{R object to serialize.}

\item{dirname}{a directory above both 1) the future rds file and 2) the R file.}

\item{rpath}{relative to \code{dirname}, the path to the future rds file.}
}
\description{
Write rds files.
}
\seealso{
\code{\link{import_rds}}
}
