% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tsv.R
\name{export_tsv}
\alias{export_tsv}
\title{Export tab-separated-values (tsv) files}
\usage{
export_tsv(tbl, path)
}
\arguments{
\item{tbl}{A table.}

\item{path}{A path.}
}
\value{

}
\description{
Execute readr::write_tsv with assumptions about values of some arguments.
}
\details{
Only blank cells represent missing data.
Over-write existing file if one exists.
Write column names at top of file.
}
\keyword{}
