% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrn.R
\name{fmrn}
\alias{fmrn}
\title{Format medical record numbers (MRNs)}
\usage{
fmrn(x, lz)
}
\arguments{
\item{x}{an integer, real, or character vector.}

\item{lz}{logical: should leading zeros be added? FALSE subtracts leading zeros.}
}
\description{
Returns a character vector containing medical record numbers (MRNs) that are with or without leading zeros.
}
\details{
With leading zeros, MRNs are 9 digits.
When searching for patients by MRN, Provide Enterprise requires 9-digit MRNs.
}
\examples{
fmrn(x = 1, lz = TRUE)
fmrn(x = 000000001, lz = FALSE)
}
