% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_sub.R
\name{g_sub}
\alias{g_sub}
\title{g_sub}
\usage{
g_sub(x, pattern, replacement, ignore.case)
}
\arguments{
\item{x}{something}

\item{pattern}{something else}

\item{replacement}{something something}

\item{ignore.case}{something something else}
}
\value{
g_sub
}
\description{
g_sub
}
\details{
g_sub
}
\examples{

g_sub(x = "hello", pattern = "HELLO", replacement = "", ignore.case = TRUE)

}
