% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_l.R
\name{grep_l}
\alias{grep_l}
\title{grep_l}
\usage{
grep_l(x, pattern, ignore.case)
}
\arguments{
\item{x}{something}

\item{pattern}{something else}

\item{ignore.case}{something something else}
}
\value{
grep_l.
}
\description{
grep_l
}
\details{
grep_l
}
\examples{

grep_l(x = "hello", pattern = "HELLO", ignore.case = TRUE)

}
