% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_df.R
\name{i_df}
\alias{i_df}
\title{Character case}
\usage{
i_df(x)
}
\arguments{
\item{x}{a data-frame.}
}
\value{
A data-frame.
}
\description{
Translate characters from upper to lower case. Leave non-alphabetic characters un-changed.
}
\examples{
ToothGrowth$supp <- as.character(x = ToothGrowth$supp)
zhaoy::i_df(x = ToothGrowth)
}
