% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_df.R
\name{import_df}
\alias{import_df}
\title{Read data-frames.}
\usage{
import_df(x)
}
\arguments{
\item{range}{A data-frame.}
}
\value{
A base-R data-frame.
}
\description{
Converts column names and categorical data to lower-case.
}
\seealso{
\code{\link{import_excel import_feather}}
}
