% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Read xls and xlsx files.}
\usage{
import_excel(criterion, file, sheet, range, guess_max)
}
\arguments{
\item{criterion}{\code{criterion} for \code{\link{find_root}}.}

\item{file}{Name and extension of the xls / xlsx file.}

\item{sheet}{Sheet to read.}

\item{range}{A cell range to read from.}

\item{guess_max}{Maximum number of data rows to use for guessing column types.}
}
\value{
A tibble.
}
\description{
Execute readxl::read_excel with assumptions about the values of some arguments.
}
\details{
\code{col_names}: Use the first row as column names.

\code{col_types}: Guess all column types from the spread-sheet.

\code{na}: Treat blank cells as missing data.

\code{trim_ws}: Trim leading and trailing whitespace.

\code{skip}: The minimum number of rows to skip before reading anything, is 0.

\code{n_max}: The maximum number of data rows to read is infinite.

Convert column names to lower-case.
}
\examples{
ds <- readxl_example(path = "datasets.xlsx")
dir <- g_sub(x = ds, pattern = "/datasets.xlsx", replacement = "", ignore.case = TRUE)
file <- g_sub(x = ds, pattern = dir, replacement = "", ignore.case = TRUE)
setwd(dir = dir)
import_excel(criterion = file, file = file, sheet = "iris", range = NULL, guess_max = 10)
}
\seealso{
\code{\link{import_tsv}}
}
