% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_feather.R
\name{import_feather}
\alias{import_feather}
\title{Read Feather files.}
\usage{
import_feather(folder, path)
}
\arguments{
\item{folder}{Any folder above both 1) the Feather file and 2) the R file.}

\item{path}{Relative to \code{folder}, path to the Feather file.}
}
\value{
A base-R data-frame.
}
\description{
Combines \code{feather::read_feather} and \code{rprojroot::find_root}.

Converts upper-case variable names and categorical data to lower-case.
}
\seealso{
\code{\link{import_df}}
\code{\link{import_excel}}
}
