% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rds.R
\name{import_rds}
\alias{import_rds}
\title{Read rds files}
\usage{
import_rds(dirname, rpath)
}
\arguments{
\item{dirname}{a directory above both 1) the rds file and 2) the R file.}

\item{rpath}{relative to \code{dirname}, the path to the rds file.}
}
\value{
A tibble.
}
\description{
Read rds files. Translate upper-case alphabetic characters to lower-case.
}
\seealso{
\code{\link{export_rds}}
}
