% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tsv.R
\name{import_tsv}
\alias{import_tsv}
\title{Read a tsv into a data-frame.}
\usage{
\code{import_tsv(criterion, file, guess_max = 10)}
}
\arguments{
\item{criterion}{\code{criterion} argument of rprojroot::\code{\link{find_root}}.}

\item{file}{Path to a file, excluding the root directory.}

\item{guess_max}{Maximum number of records to use for guessing column types.}
}
\value{
A data-frame.
}
\description{
Execute readr::\code{\link{read_tsv}} with pre-set values in some arguments.
Convert column names and categorical data to lower-case.
}
\details{
The readr::\code{\link{read_tsv}} pre-set argument values are

\code{quote = ""}: Use nothing to quote strings.

\code{colnames = TRUE}: Use the first row of the input as the column names.

\code{col_types = NULL}: Impute all column types from the first 1000 rows on the input.

\code{locale = default_locale()}: Use readr::\code{\link{default_locale}} as the locale.

\code{na = ""}: Treat blank cells as missing values.

\code{quoted_na = TRUE}: Treat missing values inside quotes as missing values.

\code{comment = ""}: Use nothing to identify comments.

\code{trim_ws = TRUE}: Trim leading and trailing white-space from each field before parsing it.

\code{skip = 0}: The number of lines to skip before reading data is 0.

\code{n_max = Inf}: The maximum number of records to read is \code{\link{Inf}}.

\code{progress = TRUE}: Display a progress bar.
}
\examples{
path_package <- path.package(package = "zhaoy", quiet = FALSE)
mtcarz <- file.path(path_package, "data/mtcars.tsv", fsep = "/")
dir <- dirname(path = mtcarz)
base <- basename(path = mtcarz)
setwd(dir = dir)
import_tsv(criterion = base, file = base, guess_max = 10)
}
\seealso{
\code{\link{import_excel}}
}
