% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_sf_numeric.R
\name{internal_sf_numeric}
\alias{internal_sf_numeric}
\title{Convert corrupted scientific-notation characters to fixed-notation numbers}
\usage{
internal_sf_numeric(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A numeric vector.
}
\description{
Convert corrupted scientific-notation characters to fixed-notation numbers.
}
