% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_df.R
\name{lc_df}
\alias{lc_df}
\title{Character Translation}
\usage{
lc_df(x)
}
\arguments{
\item{x}{a data-frame.}
}
\value{
A tibble.
}
\description{
Translate characters from upper to lower case.
}
\details{
Non-alphabetic characters and non-character vectors are left un-changed.
}
\examples{
ToothGrowth$supp <- as.character(x = ToothGrowth$supp)
lc_df(x = ToothGrowth)
}
