% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_convert.R
\name{left_convert}
\alias{left_convert}
\title{Convert data via left join}
\usage{
left_convert(category, x_df, x_var, y_df, y_var)
}
\arguments{
\item{category}{Category.}

\item{x_df}{Left-hand-side data-frame.}

\item{x_var}{Left-hand-side variable.}

\item{y_df}{Left-hand-side data-frame.}

\item{y_var}{Left-hand-side variable.}
}
\value{
A tibble.
}
\description{
Convert data via left join.
}
