% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_convert_helper.R
\name{left_convert_helper}
\alias{left_convert_helper}
\title{Convert data by left-joining data-frames and look-up tables}
\usage{
left_convert_helper(x_df, x_var, y_df, category)
}
\arguments{
\item{x_df}{Left-side data-frame.}

\item{x_var}{Left-side variable, must be character vector while using this function.}

\item{y_df}{Right-side data-frame.}

\item{category}{Right-side data-frame's category.}
}
\value{
A character vector.
}
\description{
Convert data by left-joining a data-frame and a look-up table.
}
