% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrn.R
\name{mrn}
\alias{mrn}
\title{Add or subtract leading zeros from medical record numbers (MRNs).}
\usage{
mrn(x, lz)
}
\arguments{
\item{x}{An object containing MRNs.}

\item{lz}{logical: should leading zeros be added? FALSE subtracts leading zeros.}
}
\value{

}
\description{
Add or subtract leading zeros from medical record numbers (MRNs).
}
\details{
With leading zeros, MRNs are 9 digits.
When searching for patients by MRN, Provide Enterprise requires 9-digit MRNs.
}
\examples{
mrn(x = 1, lz = TRUE)
mrn(x = 000000001, lz = FALSE)
}
\keyword{}
