% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_missing.R
\name{n_missing}
\alias{n_missing}
\title{n_missing}
\usage{
n_missing(x)
}
\arguments{
\item{x}{something}
}
\value{
n_missing
}
\description{
n_missing
}
\details{
n_missing
}
\examples{

n_missing(x = mtcars)

}
