% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_unique.R
\name{order_unique}
\alias{order_unique}
\title{Extract and order unique elements of vectors}
\usage{
order_unique(x)
}
\arguments{
\item{x}{an R object with a class or a numeric, complex, character, or logical vector.}
}
\description{
Order unique elements of vectors or factors into ascending order, beginning with any NAs.
}
\examples{
order_unique(x = iris$Sepal.Length)
order_unique(x = mtcars$mpg)
}
