% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_date.R
\name{origin_date}
\alias{origin_date}
\title{Days-since-origin conversion to Date class}
\usage{
origin_date(x, origin)
}
\arguments{
\item{x}{An object representing days-since-origin, to be converted.}

\item{origin}{A Date object, or something which can be coerced by as.Date(origin, ...) to such an object.}
}
\description{
Function to convert representations of days-since-origin to objects of class "Date" representing dates.
Executes as.Date with assumptions about the values of some arguments.
}
\details{
format is un-necessary.

tz is the current time zone.
}
\examples{
origin_date(x = 32768, origin = "1900-1-1")
}
