% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Convert relative file-paths to absolute}
\usage{
path(dirname, rpath)
}
\arguments{
\item{dirname}{A directory above 1) the file / folder and 2) the R file.}

\item{rpath}{Relative to \code{dirname}, path to the file / folder.}
}
\value{
A character vector.
}
\description{
Convert relative file-paths to absolute.
}
