% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_lab.R
\name{pe_lab}
\alias{pe_lab}
\title{Provide Enterprise lab}
\usage{
pe_lab(criterion, file, sheet)
}
\arguments{
\item{criterion}{criterion}

\item{file}{file}

\item{sheet}{sheet}
}
\description{
Provide Enterprise lab
}
