% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_expr.R
\name{reg_expr}
\alias{reg_expr}
\title{reg_expr}
\usage{
reg_expr(x, pattern, ignore.case)
}
\arguments{
\item{x}{something}

\item{pattern}{something else}

\item{ignore.case}{something something else}
}
\value{
reg_expr
}
\description{
reg_expr
}
\details{
reg_expr
}
\examples{

reg_expr(x = "hello", pattern = "HELLO", ignore.case = TRUE)

}
