% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_matches.R
\name{reg_matches}
\alias{reg_matches}
\title{reg_matches}
\usage{
reg_matches(x, m)
}
\arguments{
\item{x}{something}

\item{m}{something else}
}
\value{
reg_matches
}
\description{
reg_matches
}
\details{
reg_matches
}
\examples{

reg_matches(x = "hello", m = reg_expr(x = "hello", pattern = "HELLO", ignore.case = TRUE))

}
