% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_col.R
\name{s_col}
\alias{s_col}
\title{Column-oriented counts and percents.}
\usage{
s_col(x)
}
\arguments{
\item{x}{a data-frame column.}
}
\value{
A base-R data-frame.
The "value" column is in ascending order, beginning with any \code{\link{NA}}s.
The "pct" column rounds numbers to the nearest integers.
}
\description{
Tabulate data as counts and percents of data-frame columns.
}
\examples{
s_col(x = attenu$station)

}
\seealso{
\code{\link{s_mode} \link{s_tbl}}
}
