% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_cp.R
\name{s_cp}
\alias{s_cp}
\title{Counts and percents.}
\usage{
s_cp(x)
}
\arguments{
\item{x}{a vector or factor.}
}
\value{
A base-R data-frame with the following columns:

value: unique values,
in ascending order with the exception of \code{\link{NA}},
which is first if it is present.

n: counts.

pct: percents rounded to the nearest integers.
}
\description{
Tabulate counts and percents of unique values within vectors and factors.
}
\examples{
s_cp(x = attenu$station)
}
\seealso{
\code{\link{s_mode} \link{s_s}}
}
