% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_cross_n.R
\name{s_cross_n}
\alias{s_cross_n}
\title{2- or 3-way frequency tables of counts}
\usage{
s_cross_n(x, ...)
}
\arguments{
\item{x}{A data-frame.}

\item{...}{2 or 3 variables.}
}
\value{
A data-frame.
}
\description{
Cross-tabulate counts of unique values, including missing-data, in 2 or 3 variables.
}
\seealso{
\code{\link{s_cross_p} \link{s_cross_np}}
}
