% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_cross_np.R
\name{s_cross_np}
\alias{s_cross_np}
\title{2- or 3-way frequency tables}
\usage{
s_cross_np(x, ..., direction)
}
\arguments{
\item{x}{A data-frame.}

\item{...}{2 or 3 variables.}

\item{direction}{Direction in which to calculate percents: "row", "col", or "all".}
}
\value{
A data-frame.
}
\description{
Cross-tabulate counts and percents of unique values, including missing-data, in 2 or 3 variables.
}
\examples{
zhaoy::s_cross_np(x = attenu, event, mag, direction = "col")
}
\seealso{
\code{\link{s_cross_n} \link{s_cross_p}}
}
