% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_hiv.R
\name{s_hiv}
\alias{s_hiv}
\title{HIV status summaries}
\usage{
s_hiv(lab_df, lab_date, cd4, vl, med_df, med_date)
}
\arguments{
\item{lab_df}{a table of lab data.}

\item{lab_date}{a column of dates.}

\item{cd4}{a column of cd4 counts.}

\item{vl}{a column of viral loads.}

\item{med_df}{a column of drug names.}

\item{med_date}{a column of dates.}
}
\value{
A graph.
}
\description{
Graphs cd4 counts, viral loads in millions, and medication names, over time.
}
