% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mode.R
\name{s_mode}
\alias{s_mode}
\title{Statistical mode.}
\usage{
s_mode(x)
}
\arguments{
\item{x}{a vector factor, or one or more dates / date-times.}
}
\value{
A vector.

If \code{\link{NA}} is the most frequent value in \code{x}, \code{\link{NA}} is returned.

If \code{x} has a length of one, or has otherwise no mode, "no mode" is returned.
}
\description{
Calculate the statistical mode.
}
\examples{
s_mode(x = attenu$dist)
}
\seealso{
\code{\link{s_cp} \link{s_s}}
}
