% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_stratify.R
\name{s_stratify}
\alias{s_stratify}
\title{Summarize data in uni-variate or bivariate tables}
\usage{
s_stratify(data, formula = NULL, strata_2 = NULL, min_dis, max_level, ...)
}
\arguments{
\item{data}{Data-frame.}

\item{formula}{For uni-variate tables: one variable followed by "~ .". For bivariate tables: use "~" to separate dependent and independent variable(s), use "+" to separate independent variables.}

\item{strata_2}{Applies only to bivariate tables. Beyond the dependent variable in \code{formula}, additional variable to stratify on. Must be character while using this function.}

\item{min_dis}{If a non-factor dependent variable has less than \code{min.dis} distinct data-points, then this function treats the variable as categorical.}

\item{max_level}{Maximum number of distinct data-points of dependent variable, categorical independent variables, and \code{strata_2}.}

\item{...}{Additional arguments.}
}
\value{
A uni-variate or bivariate table.
}
\description{
Summarize data in uni-variate or bivariate tables.
}
