% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_tbl.R
\name{s_tbl}
\alias{s_tbl}
\title{Summary statistics of columns and tables}
\usage{
s_tbl(x)
}
\arguments{
\item{x}{a column or table}
}
\value{
A table.
}
\description{
Tabulate summary statistics of columns and tables.
}
\details{
The outputs are:

col: column name, displayed only when input is a table

n_miss: number of missing data

pct_miss: percent of column that are missing data

n_unique: number of unique values in column

pct_unique: percent of column that are unique values

min, max, median, mean: \code{\link{NA}} when a column is non-numeric

mode: \code{\link{NA}} only when \code{\link{NA}} is the most frequent value in a column.
If there is no mode, the displayed value is "no mode".
If there is more than one mode, the displayed value is "> 1 mode".
}
\examples{
s_tbl(x = attenu$station)
s_tbl(x = attenu)
}
\seealso{
\code{\link{s_col} \link{s_mode}}
}
