% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_unique.R
\name{s_unique}
\alias{s_unique}
\title{Unique Values}
\usage{
s_unique(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
A data-frame with the following variables:

value: value name.

n: count of value.

pct: percent of \code{x} that has the value, rounded to one decimal place.
}
\description{
Tabulate counts and percents of unique values, including \code{NA}.
}
\examples{
s_unique(x = attenu$station)
}
\seealso{
\code{\link{s_mode} \link{s_s}}
}
