% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_numeric.R
\name{sf_numeric}
\alias{sf_numeric}
\title{Convert corrupted scientific-notation characters to fixed-notation numbers}
\usage{
sf_numeric(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A numeric vector.
}
\description{
Convert corrupted scientific-notation characters to fixed-notation numbers.
}
\examples{
sf_numeric(x = "1+003")
}
