% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_unique.R
\name{sort_unique}
\alias{sort_unique}
\title{Sort unique values of vectors}
\usage{
sort_unique(x)
}
\arguments{
\item{x}{A vector.}
}
\value{

}
\description{
Sort unique values of vectors, in ascending order beginning with any NAs.
}
\details{
Values can be categories, numbers, or strings.
}
\examples{
sort_unique(x = iris$Sepal.Length)
sort_unique(x = mtcars$mpg)
}
\keyword{}
