% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squish_df.R
\name{squish_df}
\alias{squish_df}
\title{Trim excess space from strings}
\usage{
squish_df(x)
}
\arguments{
\item{x}{A data-frame.}
}
\value{
A tibble.
}
\description{
Trim excess space from strings.
}
\examples{
ToothGrowth$supp <- as.character(x = ToothGrowth$supp)
zhaoy::squish_df(x = ToothGrowth)
}
