% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_vector.R
\name{summarize_vector}
\alias{summarize_vector}
\title{Summarize values of vectors}
\usage{
\code{summarize_vector(x)}
}
\arguments{
\item{x}{a numeric, complex, character, or logical vector.}
}
\description{
Display 1) values of vectors in ascending order, beginning with any NAs,
2) counts, and 3) percentages.
}
\examples{
summarize_vector(x = iris$Sepal.Length)
summarize_vector(x = mtcars$mpg)
}
