% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_activity.R
\name{tidy_activity}
\alias{tidy_activity}
\title{Tidy Provide Enterprise activity data.}
\usage{
tidy_activity(x)
}
\arguments{
\item{x}{A data-frame.}
}
\value{
A tibble.
}
\description{
Tidy the Provide Enterprise "Activity Summary By Provider by Client by Date" report.
}
