% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_adap.R
\name{tidy_adap}
\alias{tidy_adap}
\title{Tidy Provide Enterprise ADAP data.}
\usage{
tidy_adap(x)
}
\arguments{
\item{x}{A data-frame.}
}
\value{
A tibble.
}
\description{
Tidy the Provide Enterprise "Clients - View ADAP Clients by Recertification Date" report.
}
