% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pea.R
\name{tidy_pea}
\alias{tidy_pea}
\title{Tidy Provide Enterprise activity data.}
\usage{
tidy_pea(x)
}
\arguments{
\item{x}{a data-frame.}
}
\value{
A data-frame.
}
\description{
Tidy the .xlsx version of the Provide Enterprise "Activity Summary by Provider by Client by Date" report.
}
