% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pel.R
\name{tidy_pel}
\alias{tidy_pel}
\title{Tidy Provide Enterprise laboratory data.}
\usage{
tidy_pel(x)
}
\arguments{
\item{x}{a data-frame.}
}
\value{
A data-frame.
}
\description{
Tidy the .xlsx version of the Provide Enterprise "Test Results by Client with ID" report.
}
