% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_provider_relationship.R
\name{tidy_provider_relationship}
\alias{tidy_provider_relationship}
\title{Tidy xovide Enterxise provider-relationship data.}
\usage{
tidy_provider_relationship(x)
}
\arguments{
\item{x}{A data-frame.}
}
\value{
A tibble.
}
\description{
Tidy the xovide Enterxise "Case Loads by provider Relationship" report.
}
