% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_lc_df.R
\name{zhaoy_lc_df}
\alias{zhaoy_lc_df}
\title{Convert upper-case English characters to lower-case}
\usage{
zhaoy_lc_df(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A vector of the same length as \code{x}.
}
\description{
Convert upper-case English characters to lower-case.
}
