% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_mode.R
\name{zhaoy_mode}
\alias{zhaoy_mode}
\title{Statistical mode.}
\usage{
zhaoy_mode(x)
}
\arguments{
\item{x}{a character, factor, logical, or numeric vector.}
}
\value{
A character, logical, or numeric vector of length one.
When the data has multiple modes, the return value is "> 1 mode".
When the data has no mode, the return value is "no mode".
}
\description{
Calculate the statistical mode.
}
\examples{
zhaoy_mode(x = attenu$dist)
}
