% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_mode.R
\name{zhaoy_s_mode}
\alias{zhaoy_s_mode}
\title{Statistical mode.}
\usage{
zhaoy_s_mode(x)
}
\arguments{
\item{x}{a numeric or logical vector, or a date / date-time / time-interval object.}
}
\value{
A length-one character vector.

If multiple modes exist, "s_mode()" is returned.

If the mode is \code{\link{NA}}, \code{\link{NA}} is returned.

If no mode exists, "no mode" is returned.
}
\description{
Calculate the statistical mode.
}
