% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_numeric.R
\name{zhaoy_s_numeric}
\alias{zhaoy_s_numeric}
\title{Summary statistics of numeric data.}
\usage{
zhaoy_s_numeric(x, fun)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A numeric vector of length one.
For non-numeric data, the return value is \code{\link{NA}}.
}
\description{
For numeric data, calculate the minimum, maximum, median, or mean.
}
\examples{
zhaoy_s_numeric(x = attenu$accel, fun = "min")
}
