% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_s.R
\name{zhaoy_s_s}
\alias{zhaoy_s_s}
\title{Summary statistics.}
\usage{
zhaoy_s_s(x, s)
}
\arguments{
\item{x}{a numeric or logical vector, or a date / date-time / time-interval object.}

\item{s}{a summary statistic.}
}
\value{
A length-one vector.

If \code{x} is non-numeric and not a date / date-time / time-interval object, \code{\link{NA}} is returned.

If \code{x} is a date / date-time / time-interval object and the summary statistic is median or mean, \code{\link{NA}} is returned.
}
\description{
Calculate the minimum, maximum, median, or mean.
}
