% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_tolower.R
\name{zhaoy_tolower}
\alias{zhaoy_tolower}
\title{Convert only character or factor data to lower-case.}
\usage{
zhaoy_tolower(x)
}
\arguments{
\item{x}{a character or factor vector.}
}
\value{
A character or factor vector.
}
\description{
Execute \code{\link{tolower}} if data is \code{\link{character}} or \code{\link{factor}}.
}
