% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc}
\alias{dc}
\title{Date-time conversion to character}
\usage{
dc(x)
}
\arguments{
\item{x}{An object in year-month-day format, to be converted.}
}
\value{
Dates of class "\code{character}", in year-month-day format.
}
\description{
Function to convert non-character representations to objects of class "\code{character}" representing dates.
Executes \code{strftime} with assumptions about the values of some arguments.
}
\examples{
# read in date info in format "yyyy-mm-dd"
x <- c("1960-1-1", "1960-1-2", "1960-3-31", "1960-7-30")
z <- dc(x = x)
z

}
\seealso{
\code{\link{dd}}
}
