% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{dd}
\alias{dd}
\title{Date conversion to Date class}
\usage{
dd(x)
}
\arguments{
\item{x}{An object in year-month-day format, to be converted.}
}
\description{
Function to convert non-"Date" representations of dates and date-times to objects of class "Date" representing dates.
Executes as.Date with assumptions about the values of some arguments.
}
\details{
In x, leading zeros in single-digit months or days, are optional.

format is "%Y-%m-%d".

origin is blank.

tz is the current time zone.
}
\examples{
x <- c("1960-1-1", "1960-1-2", "1960-3-31", "1960-7-30")
z <- dd(x = x)
z
}
