% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{export_excel}
\alias{export_excel}
\title{Export to xlsx}
\usage{
export_excel(x, dirname, rpath)
}
\arguments{
\item{x}{data-frame, or named list of data-frames that are sheets in the future xlsx file.}

\item{dirname}{a directory above both 1) the future xlsx file and 2) the R file.}

\item{rpath}{relative to \code{dirname}, the path to the future xlsx file.}
}
\description{
Write a data-frame to an xlsx file. Support strings, numbers, booleans, and dates.
}
\seealso{
\code{\link{import_excel}}
}
