% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_feather.R
\name{export_feather}
\alias{export_feather}
\title{Write Feather files.}
\usage{
export_feather(x, folder, path)
}
\arguments{
\item{x}{A data-frame to write to disk.}

\item{folder}{Any folder above both 1) the Feather file and 2) the code file.}

\item{path}{Relative to \code{folder}, path to the Feather file.}
}
\value{
A data-frame.
}
\description{
Combines \code{feather::write_feather} and \code{rprojroot::find_root}.
}
