% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_feather.R
\name{export_feather}
\alias{export_feather}
\title{Export data-frames as Feather files}
\usage{
export_feather(x, dirname, rpath)
}
\arguments{
\item{x}{A data-frame.}

\item{dirname}{A directory above 1) the Feather file and 2) the R file.}

\item{rpath}{Relative to \code{dirname}, path to the Feather file.}
}
\description{
Export data-frames as Feather files.
}
\seealso{
\code{\link{import_feather}}
}
