% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_feather.R
\name{export_feather}
\alias{export_feather}
\title{Write Feather files.}
\usage{
export_feather(x, folder, path)
}
\arguments{
\item{x}{A data-frame to write to disk.}

\item{folder}{Any folder above both 1) the Feather file and 2) the file that contains the code.}

\item{path}{Path to the Feather file, excluding \code{folder}.}
}
\value{
A base-R data-frame.
}
\description{
Executes feather::\code{\link{write_feather}}.
}
