% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_feather.R
\name{export_feather}
\alias{export_feather}
\title{Write Feather files.}
\usage{
export_feather(x, folder, path)
}
\arguments{
\item{x}{a data-frame to write to disk.}

\item{folder}{any folder above the R file.}

\item{path}{relative to \code{folder}, path to Feather file.}
}
\description{
Write Feather files.
}
\seealso{
\code{\link{import_feather}}
}
