% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_feather.R
\name{export_feather}
\alias{export_feather}
\title{Write data-frames to Feather files.}
\usage{
export_feather(x, folder, path)
}
\arguments{
\item{x}{A data-frame to write to disk.}

\item{folder}{Any folder above both 1) the path to write to and 2) the R file.}

\item{path}{Relative to \code{folder}, path to write to.}
}
\description{
Combines \code{feather::write_feather} and \code{rprojroot::find_root}.
}
