% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tsv.R
\name{export_tsv}
\alias{export_tsv}
\title{Write a data frame to a tab-separated-values (tsv) file}
\usage{
export_tsv(x, folder)
}
\arguments{
\item{x}{A data frame to write to disk.}

\item{folder}{Path or connection to folder.}
}
\description{
Never writes row names. Executes readr::write_tsv with assumptions about the values of some arguments.
}
\details{
NA is used for missing values.

Over-write existing file.

Write column names at the top of the file.
}
\examples{
dir <- getwd()
export_tsv(x = mtcars, folder = dir)
}
