% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tsv.R
\name{export_tsv}
\alias{export_tsv}
\title{Write a data-frame to a tsv file.}
\usage{
export_tsv(x, folder)
}
\arguments{
\item{x}{A data-frame to write to disk.}

\item{folder}{Path or connection to the folder that will store the data-frame.}
}
\value{
\code{export_*()} returns the input x invisibly.
}
\description{
Execute readr::\code{\link{write_tsv}} with pre-set values in some arguments.
}
\details{
The readr::\code{\link{write_tsv}} pre-set argument values are

\code{na = ""}: Use blanks for missing values.

\code{append = FALSE}: Over-write existing file.

\code{col_names = TRUE}: Write column names at the top of the file.
}
\examples{
dir <- getwd()
export_tsv(x = mtcars, folder = dir)
}
