% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path.R
\name{file_path}
\alias{file_path}
\title{Build absolute paths to files or folders}
\usage{
file_path(dirname, rpath)
}
\arguments{
\item{dirname}{a directory above 1) the file / folder and 2) the R file.}

\item{rpath}{relative to \code{dirname}, path to the file / folder.}
}
\value{
A character vector.
}
\description{
Build absolute paths to files or folders.
}
