% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_excel.R
\name{i_excel}
\alias{i_excel}
\title{Import Microsoft Excel files}
\usage{
i_excel(folder, path, sheet = NULL, range = NULL)
}
\arguments{
\item{folder}{a folder above 1) the Excel file and 2) the R file.}

\item{path}{relative to \code{folder}, path to the Excel file.}

\item{sheet}{optional: a sheet to import.}

\item{range}{optional: a cell range to import.}
}
\value{
A data-frame.
}
\description{
Import Microsoft Excel files.

Convert upper-case English characters to lower-case.
}
