% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_feather.R
\name{i_feather}
\alias{i_feather}
\title{Read Feather files}
\usage{
i_feather(folder, path)
}
\arguments{
\item{folder}{a folder above both 1) the Feather file and 2) the R file.}

\item{path}{relative to \code{folder}, the path to the Feather file.}
}
\value{
A data-frame.
}
\description{
Read Feather files. Translate upper-case alphabetic characters to lower-case.
}
