% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_feather.R
\name{i_feather}
\alias{i_feather}
\title{Import Feather files}
\usage{
i_feather(folder, path)
}
\arguments{
\item{folder}{a folder above 1) the Feather file and 2) the R file.}

\item{path}{relative to \code{folder}, path to the Feather file.}
}
\value{
A data-frame.
}
\description{
Import Feather files.

Convert upper-case English characters to lower-case.
}
