% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_df.R
\name{import_df}
\alias{import_df}
\title{Convert data-frame characters to lower-case.}
\usage{
import_df(x)
}
\arguments{
\item{x}{a data-frame.}
}
\value{
A data-frame.
}
\description{
Convert characters from upper to lower case, in variable names and in data points.
}
\examples{
ToothGrowth$supp <- as.character(x = ToothGrowth$supp)
import_df(x = ToothGrowth)
}
