% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_df.R
\name{import_df}
\alias{import_df}
\title{Read data-frames.}
\usage{
import_df(x)
}
\arguments{
\item{x}{a data-frame.}
}
\value{
A base-R data-frame.
}
\description{
Convert upper-case variable names and categorical data to lower-case.
}
\examples{
import_df(x = iris)
}
