% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Read xls and xlsx files.}
\usage{
\code{import_excel(criterion, path, sheet = NULL, range = NULL, guess_max = 10)}
}
\arguments{
\item{criterion}{\code{criterion} argument of rprojroot::\code{\link{find_root}}.}

\item{path}{Path to the xls / xlsx file, excluding the root directory.}

\item{sheet}{Sheet to read.}

\item{range}{A cell range to read from.}

\item{guess_max}{Maximum number of data rows to use for guessing column types.}
}
\value{
A data-frame.
}
\description{
Execute readxl::\code{\link{read_excel}} with pre-set values in some arguments.
Convert column names and categorical data to lower-case.
}
\details{
The readxl::\code{\link{read_excel}} pre-set argument values are

\code{col_names = TRUE}: Use the first row as column names.

\code{col_types = NULL}: Guess all column types from the spread-sheet.

\code{na = ""}: Treat blank cells as missing data.

\code{trim_ws = TRUE}: Trim leading and trailing white-space.

\code{skip = 0}: The minimum number of rows to skip before reading anything is 0.

\code{n_max = Inf}: The maximum number of data rows to read is \code{\link{Inf}}.
}
\examples{
library(package = readxl)
datasets <- readxl_example(path = "datasets.xlsx")
dir <- dirname(path = datasets)
base <- basename(path = datasets)
setwd(dir = dir)
import_excel(criterion = base, path = base, guess_max = 10)
}
\seealso{
\code{\link{import_tsv}}
}
