% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Import Microsoft Excel files}
\usage{
import_excel(file, sheet, range, guess_max)
}
\arguments{
\item{file}{A file.}

\item{sheet}{A sheet.}

\item{range}{A cell range to read from.}

\item{guess_max}{Maximum number of rows to use for guessing column types.}
}
\value{

}
\description{
Execute readxl::read_excel with assumptions about the values of some arguments.
}
\details{
Range takes precedence over skip, n_max, and sheet. If Range is un-necessary, set it to NULL.
First rows are column names.
Guess column types from the spread-sheet.
Only blank cells represent missing data.
Trim leading and trailing white-space.
There are no rows to skip.
There is no upper bound on the number of rows to read.
}
\keyword{}
