% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Read xls and xlsx files}
\usage{
import_excel(dirname, rpath, sheet = NULL, range = NULL)
}
\arguments{
\item{dirname}{a directory above both 1) the xls / xlsx file and 2) the R file.}

\item{rpath}{relative to \code{dirname}, the path to the xls / xlsx file.}

\item{sheet}{(optional) sheet to read.}

\item{range}{(optional) a cell range to read from.}
}
\value{
A tibble.
}
\description{
Read xls and xlsx files. Translate upper-case alphabetic characters to lower-case.
}
\seealso{
\code{\link{export_excel}}
}
