% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Read xls and xlsx files.}
\usage{
import_excel(folder, path, sheet = NULL, range = NULL, guess_max)
}
\arguments{
\item{folder}{Folder above the xls / xlsx file.}

\item{path}{Path to the file, excluding \code{folder}.}

\item{sheet}{Sheet to read.}

\item{range}{A cell range to read from.}

\item{guess_max}{Maximum number of data rows to use for guessing column types.}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Executes readxl::\code{\link{read_excel}} with pre-set values in some arguments.
Converts column names and categorical data to lower-case.
}
\details{
The pre-set values in some arguments of readxl::\code{\link{read_excel}}, are:

\code{col_names = TRUE}: Use the first row as column names.

\code{col_types = NULL}: Guess all column types from the spread-sheet.

\code{na = ""}: Treat blank cells as missing data.

\code{trim_ws = TRUE}: Trim leading and trailing white-space.

\code{skip = 0}: Skip a minimum of 0 rows before reading anything.

\code{n_max = Inf}: Read a maximum of \code{\link{Inf}} rows.
}
