% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Import Microsoft Excel files}
\usage{
import_excel(dirname, rpath, sheet = NULL, range = NULL)
}
\arguments{
\item{dirname}{A directory above 1) the Excel file and 2) the R file.}

\item{rpath}{Relative to \code{dirname}, path to the Excel file.}

\item{sheet}{Optional: a sheet to import.}

\item{range}{Optional: a cell range to import.}
}
\value{
A tibble.
}
\description{
Import Microsoft Excel files.

Convert upper-case English characters to lower-case.
}
\seealso{
\code{\link{export_excel}}
}
