% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Read xls and xlsx files.}
\usage{
import_excel(folder, path, sheet = NULL, range = NULL)
}
\arguments{
\item{folder}{Any folder above both 1) the xls / xlsx file and 2) the file that contains the code.}

\item{path}{Path to the xls / xlsx file, excluding \code{folder}.}

\item{sheet}{Sheet to read.}

\item{range}{A cell range to read from.}
}
\value{
A base-R data-frame.
}
\description{
Executes readxl::\code{\link{read_excel}} with pre-set values in some arguments.
Uses relative file-paths.
Converts column names and categorical data to lower-case.
}
\details{
\code{import_excel} excutes readxl::\code{\link{read_excel}} with the following pre-set argument values:

\code{col_names = TRUE}: Use the first row as column names.

\code{col_types = NULL}: Guess all from the spread-sheet.

\code{na = ""}: Treat blank cells as missing data.

\code{trim_ws = TRUE}: Trim leading and trailing white-space.

\code{skip = 0}: Skip a minimum of 0 rows before reading anything.

\code{n_max = Inf}: Read a maximum of data rows.

\code{guess_max = 100000}: Use a maximum of 100000 data rows to guess column types.
}
\seealso{
\code{\link{import_df import_feather}}
}
