% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Import Microsoft Excel files}
\usage{
import_excel(path, sheet, range, guess_max)
}
\arguments{
\item{path}{A path.}

\item{sheet}{A sheet.}

\item{range}{A cell range to read from. Range takes precedence over skip, n_max, and sheet. If Range is un-necessary, set it to NULL.}

\item{guess_max}{Maximum number of rows to use for guessing column types.}
}
\value{

}
\description{
Execute readxl::read_excel with assumptions about values of some arguments.
}
\details{
First rows are column names.
Only blank cells represent missing data.
There are no rows to skip.
}
\keyword{}
