% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Read xls and xlsx files.}
\usage{
import_excel(folder, path, sheet = NULL, range = NULL)
}
\arguments{
\item{folder}{any folder above both 1) the xls / xlsx file and 2) the R file.}

\item{path}{relative to \code{folder}, path to the xls / xlsx file.}

\item{sheet}{sheet to read.}

\item{range}{a cell range to read from.}
}
\value{
A base-R data-frame.
}
\description{
Read xls and xlsx files.

Convert upper-case variable names and categorical data to lower-case.
}
