% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_feather.R
\name{import_feather}
\alias{import_feather}
\title{Read Feather files.}
\usage{
import_feather(folder, path)
}
\arguments{
\item{folder}{any folder above both 1) the Feather file and 2) the R file.}

\item{path}{relative to \code{folder}, path to the Feather file.}
}
\value{
A base-R data-frame.
}
\description{
Read Feather files.
}
\seealso{
\code{\link{export_feather}}
}
