% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_feather.R
\name{import_feather}
\alias{import_feather}
\title{Read Feather files.}
\usage{
import_feather(folder, path)
}
\arguments{
\item{folder}{Any folder above both 1) the Feather file and 2) the file that contains the code.}

\item{path}{Path to the Feather file, excluding \code{folder}.}
}
\value{
A base-R data-frame.
}
\description{
Executes feather::\code{\link{read_feather}}.
Uses relative file-paths.
Converts column names and categorical data to lower-case.
}
\seealso{
\code{\link{import_df import_excel}}
}
