% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rds.R
\name{import_rds}
\alias{import_rds}
\title{Import rds files}
\usage{
import_rds(dirname, rpath)
}
\arguments{
\item{dirname}{A directory above 1) the rds file and 2) the R file.}

\item{rpath}{Relative to \code{dirname}, path to the rds file.}
}
\value{
A tibble.
}
\description{
Import rds files.

Convert upper-case English characters to lower-case.
}
\seealso{
\code{\link{export_rds}}
}
