% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tsv.R
\name{import_tsv}
\alias{import_tsv}
\title{Read a tsv into a tibble.}
\usage{
\code{import_tsv(criterion, file, guess_max = 10)}
}
\arguments{
\item{criterion}{\code{criterion} in rprojroot::\code{\link{find_root}}.}

\item{file}{Path to a file, excluding the root directory.}

\item{guess_max}{Maximum number of records to use for guessing column types.}
}
\value{
A data-frame.
}
\description{
Execute readr::\code{\link{read_tsv}} with pre-set values in some arguments.
Convert column names to lower-case.
}
\details{
The readr::\code{\link{read_tsv}} pre-set argument values are

\code{quote = ""}: Use blanks to quote strings.

\code{colnames = TRUE}: Use the first row of the input as the column names.

\code{col_types = NULL}: Impute all column types from the first 1000 rows on the input.

\code{locale = default_locale()}: Use readr::\code{\link{default_locale}} as the locale.

\code{na = ""}: Use blanks for missing values.

\code{quoted_na = FALSE}: Treat missing values inside quotes as strings.

\code{comment = ""}: Use blanks to identify comments.

\code{trim_ws = TRUE}: Trim leading and trailing white-space from each field before parsing it.

\code{skip = 0}: Skip 0 lines before reading data.

\code{n_max = Inf}: Set the maximum number of records to read to \code{\link{Inf}}.

\code{progress = TRUE}: Display a progress bar.
}
\seealso{
\code{\link{import_excel}}
}
