% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tsv.R
\name{import_tsv}
\alias{import_tsv}
\title{Import tab-separated-values (tsv) files}
\usage{
import_tsv(path, guess_max)
}
\arguments{
\item{path}{A path.}

\item{guess_max}{Maximum number of rows to use for guessing column types.}
}
\value{

}
\description{
Execute readr::read_tsv with assumptions about values of some arguments.
}
\details{
First rows are column names.
Guess column types.
Only blank cells represent missing data.
Missing data inside quotes constitute missing data.
There are no characters for quoting strings.
There are no strings for identifying comments.
Trim white-space.
There are no rows to skip.
There is no upper bound on the number of rows to read.
Show progress bar.
locale = default_locale()
}
\keyword{}
