% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_df.R
\name{lc_df}
\alias{lc_df}
\title{Character case}
\usage{
lc_df(x)
}
\arguments{
\item{x}{a data-frame.}
}
\value{
A tibble.
}
\description{
Translate characters from upper to lower case. Leave non-alphabetic characters un-changed.
}
\examples{
ToothGrowth$supp <- as.character(x = ToothGrowth$supp)
zhaoy::lc_df(x = ToothGrowth)
}
