% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_df.R
\name{lc_df}
\alias{lc_df}
\title{Convert upper-case English characters to lower-case}
\usage{
lc_df(x)
}
\arguments{
\item{x}{A data-frame.}
}
\value{
A tibble.
}
\description{
Convert upper-case English characters to lower-case.
}
\examples{
ToothGrowth$supp <- as.character(x = ToothGrowth$supp)
zhaoy::lc_df(x = ToothGrowth)
}
