% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz_id.R
\name{lz_id}
\alias{lz_id}
\title{Leading zeros in numeric identifiers.}
\usage{
lz_id(x, lz)
}
\arguments{
\item{x}{A vector of numeric identifiers, each up to 9 digits.}

\item{lz}{Logical: should leading zeros be included?}
}
\value{
A character vector.
}
\description{
Includes or excludes leading zeros in numeric identifiers.
}
\examples{
lz_id(x = 1, lz = TRUE)
lz_id(x = 000000001, lz = FALSE)
}
