% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz_id.R
\name{lz_id}
\alias{lz_id}
\title{Leading zeros in medical record numbers (MRNs)}
\usage{
lz_id(x, lz)
}
\arguments{
\item{x}{one or more MRNs.}

\item{lz}{logical: should leading zeros be included? FALSE excludes leading zeros.}
}
\value{
When \code{lz = TRUE}, an object of type \code{\link{character}}. \code{When lz = FALSE}, an object of type \code{\link{integer}}.
}
\description{
Includes or excludes leading zeros in MRNs.
}
\details{
With leading zeros, MRNs are 9 digits.

Provide Enterprise requires 9-digit MRNs.
}
\examples{
lz_id(x = 1, lz = TRUE)
lz_id(x = 000000001, lz = FALSE)
}
