% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz_id.R
\name{lz_id}
\alias{lz_id}
\title{Leading zeros in numeric identifiers.}
\usage{
lz_id(x, lz)
}
\arguments{
\item{x}{A vector of numeric identifiers, each up to 9 digits.}

\item{lz}{Logical: include leading zeros?}
}
\value{
A character vector.
}
\description{
Includes or excludes leading zeros in numeric identifiers.
}
\examples{
x <- c("1", "000000002")
lz_id(x = x, lz = TRUE)
lz_id(x = x, lz = FALSE)
}
