% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz_id.R
\name{lz_id}
\alias{lz_id}
\title{Leading Zeros in Epic ID Numbers}
\usage{
lz_id(x, lz)
}
\arguments{
\item{x}{at least one Epic ID number.}

\item{lz}{logical: TRUE includes, and FALSE excludes, leading zeros.}
}
\value{
A character vector.
}
\description{
Include or exclude leading zeros in Epic ID numbers.

Epic ID numbers have 1-8 positive integers and 0-8 leading zeros.
}
\examples{
x <- c("1", "000000002")
lz_id(x = x, lz = TRUE)
lz_id(x = x, lz = FALSE)
}
