% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz_id.R
\name{lz_id}
\alias{lz_id}
\title{Format MRNs with leading zeros}
\usage{
lz_id(x, lz)
}
\arguments{
\item{x}{a vector of MRNs.}

\item{lz}{logical: should leading zeros be added? FALSE excludes leading zeros.}
}
\value{
An integer or character vector.
}
\description{
Return a vector of medical record numbers (MRNs) that include or exclude leading zeros.
}
\details{
With leading zeros, MRNs are 9 digits.
Provide Enterprise requires 9-digit MRNs.
}
\examples{
lz_id(x = 1, lz = TRUE)
lz_id(x = 000000001, lz = FALSE)
}
