% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz_id.R
\name{lz_id}
\alias{lz_id}
\title{Leading zeros in integer identifiers.}
\usage{
lz_id(x, lz)
}
\arguments{
\item{x}{a character, factor, or numeric vector of up-to-9-digits integers.}

\item{lz}{logical: TRUE includes, and FALSE excludes, leading zeros.}
}
\value{
A character vector.
}
\description{
Include or exclude leading zeros in integer identifiers.
}
\examples{
x <- c("1", "000000002")
lz_id(x = x, lz = TRUE)
lz_id(x = x, lz = FALSE)
}
