% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz_id.R
\name{lz_id}
\alias{lz_id}
\title{Leading zeros in medical record numbers (MRNs)}
\usage{
lz_id(x, lz)
}
\arguments{
\item{x}{a vector of MRNS, each up to 9 digits.}

\item{lz}{logical: should leading zeros be included?.}
}
\value{
When \code{lz = TRUE}, a vector of type \code{\link{character}}.

When \code{lz = FALSE}, a vector of type \code{\link{integer}}.
}
\description{
Includes or excludes leading zeros in MRNs.
}
\details{
Provide Enterprise requires 9-digit MRNs.

MRNs that include leading zeros, are 9 digits.
}
\examples{
lz_id(x = 1, lz = TRUE)
lz_id(x = 000000001, lz = FALSE)
}
