% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrn.R
\name{mrn}
\alias{mrn}
\title{Format Medical Record Numbers (MRNs)}
\usage{
mrn(x, lz)
}
\arguments{
\item{x}{a logical, integer, real, or character vector.}

\item{lz}{logical: should leading zeros be added? FALSE subtracts leading zeros.}
}
\description{
Returns a character vector containing medical record numbers (MRNs) with or without leading zeros.
}
\details{
With leading zeros, MRNs are 9 digits.
When searching for patients by MRN, Provide Enterprise requires 9-digit MRNs.
}
\examples{
mrn(x = 1, lz = TRUE)
mrn(x = 000000001, lz = FALSE)
}
