% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Convert relative paths to absolute paths}
\usage{
path(basename, ...)
}
\arguments{
\item{basename}{Root directory.}

\item{...}{Relative to \code{basename}, path to the file or folder.}
}
\value{
A vector.
}
\description{
If the working directory is below the root directory, convert a relative path to an absolute path.
}
