% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Convert relative paths to absolute paths}
\usage{
path(basename, ...)
}
\arguments{
\item{basename}{The target folder.}

\item{...}{Relative to \code{basename}, path to the target file.}
}
\value{
A character vector.
}
\description{
If the working directory is the target folder or below, convert relative paths to absolute paths.
}
