% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Convert relative file-paths to absolute file-paths}
\usage{
path(basename, rpath)
}
\arguments{
\item{basename}{A folder above the 1) target file / folder and 2) R-code file.}

\item{rpath}{Relative to \code{basename}, path to the target file / folder.}
}
\value{
A character vector.
}
\description{
Convert relative file-paths to absolute file-paths.
}
