% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_lab.R
\name{pe_lab}
\alias{pe_lab}
\title{Read Provide Enterprise lab data}
\usage{
pe_lab(criterion, file, sheet)
}
\arguments{
\item{criterion}{A criterion.}

\item{file}{Name and extension of a file.}

\item{sheet}{Sheet to read.}
}
\description{
Read lab data that is in a .xlsx file downloaded from Provide Enterprise.
}
