% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_lab.R
\name{pe_lab}
\alias{pe_lab}
\title{Convert PE lab data into tidy data frames}
\usage{
pe_lab(criterion, file, sheet)
}
\arguments{
\item{criterion}{A criterion.}

\item{file}{Name and extension of the xlsx file.}

\item{sheet}{Sheet to read.}
}
\description{
Tidy PE lab data that were exported in xlsx files.
}
