% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_col.R
\name{s_col}
\alias{s_col}
\title{Summarize values of vectors}
\usage{
\code{s_col(x)}
}
\arguments{
\item{x}{a numeric, complex, character, or logical vector.}
}
\value{
A data-frame.
}
\description{
Tabulate 1) values of vectors in ascending order, beginning with any NAs,
2) counts, and 3) percentages.
}
\examples{
s_col(x = warpbreaks$breaks)
}
\seealso{
\code{\link{s_tbl}}
}
