% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_col.R
\name{s_col}
\alias{s_col}
\title{Column summaries}
\usage{
s_col(x)
}
\arguments{
\item{x}{a column for which a summary is desired.}
}
\value{
A table.
}
\description{
Tabulates counts and percentages of values that are in columns.
Displays the values in ascending order, beginning with any \code{\link{NA}}s.
}
\examples{
s_col(x = attenu$station)
}
\seealso{
\code{\link{s_mode} \link{s_tbl}}
}
