% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_col.R
\name{s_col}
\alias{s_col}
\title{Column-value summaries}
\usage{
s_col(x)
}
\arguments{
\item{x}{a column for which a summary of values is desired.}
}
\value{
A table.
}
\description{
Tabulate counts and percentages of column values.
Displays column values in ascending order, beginning with any NAs.
}
\examples{
s_col(x = attenu$station)
}
\seealso{
\code{\link{s_tbl}}
}
