% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_cp.R
\name{s_cp}
\alias{s_cp}
\title{Counts and percents.}
\usage{
s_cp(x)
}
\arguments{
\item{x}{a vector, factor, or one or more dates / date-time(s).}
}
\value{
A base-R data-frame with the following columns:

value: unique values,
which begin with "<NA>" if \code{\link{NA}} is present,
then continue in ascending order.

n: counts.

pct: counts as percents rounded to the nearest integers.
}
\description{
Tabulate counts and percents of unique values within vectors, factors, and dates / date-time(s).
}
\examples{
s_cp(x = attenu$station)
}
\seealso{
\code{\link{s_mode} \link{s_s}}
}
