% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_cp.R
\name{s_cp}
\alias{s_cp}
\title{Counts and percents.}
\usage{
s_cp(x)
}
\arguments{
\item{x}{a character, factor, logical, or numeric vector.}
}
\value{
A base-R data-frame with the following columns:

value: unique values, in ascending order beginning with any \code{\link{NA}}s.

n: counts.

pct: percents rounded to the nearest integers.
}
\description{
Tabulate counts and percents of vectors.
}
\examples{
s_cp(x = attenu$station)
}
\seealso{
\code{\link{s_mode} \link{s_s}}
}
