% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_cross_p.R
\name{s_cross_p}
\alias{s_cross_p}
\title{2- or 3-way frequency tables of percents}
\usage{
s_cross_p(x, ..., direction)
}
\arguments{
\item{x}{a data-frame.}

\item{...}{2 or 3 variables.}

\item{direction}{direction in which to calculate percents.}
}
\value{
A data-frame.
}
\description{
Cross-tabulate percents of unique values, including missing-data, in 2 or 3 variables.
}
\examples{
zhaoy::s_cross_p(x = attenu, event, mag, direction = "col")
}
\seealso{
\code{\link{s_cross_n} \link{s_cross_np}}
}
