% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_hiv.R
\name{s_hiv}
\alias{s_hiv}
\title{HIV status summaries}
\usage{
s_hiv(df, date, cd4, vl, med)
}
\arguments{
\item{df}{a table for which a summary is desired.}

\item{date}{a column of dates}

\item{cd4}{a column of cd4 counts}

\item{vl}{a column of viral loads}

\item{med}{a column of drug names}
}
\value{
A graph.
}
\description{
Graphs cd4 counts, viral loads in millions, and drug names, over time.
}
