% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mode.R
\name{s_mode}
\alias{s_mode}
\title{Statistical mode.}
\usage{
s_mode(x)
}
\arguments{
\item{x}{a numeric or logical vector, or a date / date-time / time-interval object.}
}
\value{
A vector.

If the mode is \code{\link{NA}}, \code{\link{NA}} is returned.

If no mode exists, "no mode" is returned.
}
\description{
Calculate the statistical mode.
}
\examples{
s_mode(x = attenu$dist)
}
\seealso{
\code{\link{s_s} \link{s_unique}}
}
