% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mode.R
\name{s_mode}
\alias{s_mode}
\title{Mathematical mode}
\usage{
s_mode(x)
}
\arguments{
\item{x}{a column.}
}
\value{
A number.
}
\description{
Calculate mode.
}
\examples{
s_mode(x = attenu$dist)
}
\seealso{
\code{\link{s_col}} \code{\link{s_tbl}}
}
