% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mode.R
\name{s_mode}
\alias{s_mode}
\title{Statistical Mode}
\usage{
s_mode(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
A vector.

If no mode exists, the result is "no mode".

If the mode is missing-data, the result is \code{NA}.
}
\description{
Compute the statistical mode.
}
\examples{
s_mode(x = attenu$dist)
}
\seealso{
\code{\link{s_s} \link{s_unique}}
}
