% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mode.R
\name{s_mode}
\alias{s_mode}
\title{Statistical mode.}
\usage{
s_mode(x)
}
\arguments{
\item{x}{a character, factor, logical, or numeric vector.}
}
\value{
A character, logical, or numeric vector.
When the data has no mode, the return value is "no mode".
}
\description{
Calculate the statistical mode.
}
\examples{
s_mode(x = attenu$dist)
}
\seealso{
\code{\link{s_cp} \link{s_s}}
}
