% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mode.R
\name{s_mode}
\alias{s_mode}
\title{Statistical mode.}
\usage{
s_mode(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
A character vector.

If the mode is \code{\link{NA}}, \code{\link{NA}} is returned.
}
\description{
Compute the statistical mode.
}
\examples{
s_mode(x = attenu$dist)
}
\seealso{
\code{\link{s_s} \link{s_unique}}
}
