% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_mode.R
\name{s_mode}
\alias{s_mode}
\title{Statistical mode}
\usage{
s_mode(x)
}
\arguments{
\item{x}{an object for which the mode is desired.}
}
\value{
An object of type \code{\link{character}} or \code{\link{numeric}}.
}
\description{
Calculates statistical mode.
}
\examples{
s_mode(x = attenu$dist)

}
\seealso{
\code{\link{s_col} \link{s_tbl}}
}
