% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_s.R
\name{s_s}
\alias{s_s}
\title{Summary statistics.}
\usage{
s_s(x)
}
\arguments{
\item{x}{a data-frame, vector, factor, or date / date-time / time-interval object.}
}
\value{
A data-frame with the following columns:

var: variable name.

n_miss: count of missing data in variable.

pct_miss: percent of variable that has missing data, rounded to one decimal place.

n_unique: count of unique values in variable.

pct_unique: percent of variable that has unique values, rounded to one decimal place.

min and max: If \code{x} is not a numeric or date / date-time / time-interval object, \code{\link{NA}} is returned.

median and mean: If \code{x} is a date / date-time / time-interval object, \code{\link{NA}} is returned.

mode:
If multiple modes exist, "s_mode()" is returned.
If the mode is \code{\link{NA}}, \code{\link{NA}} is returned.
If no mode exists, "no mode" is returned.
}
\description{
Tabulate summary statistics.
}
\examples{
s_s(x = attenu$mag)
s_s(x = attenu)
}
\seealso{
\code{\link{s_mode} \link{s_unique}}
}
