% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_s.R
\name{s_s}
\alias{s_s}
\title{Summary Statistics}
\usage{
s_s(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
A data-frame.

min and max: If \code{x} is not a numeric or date / date-time / time-interval object, \code{\link{NA}} is returned.

median and mean: If \code{x} is a date / date-time / time-interval object, \code{\link{NA}} is returned.

mode:
If the mode is \code{\link{NA}}, \code{\link{NA}} is returned.
If multiple modes exist, "s_mode()" is returned.
}
\description{
Tabulate summary statistics.
}
\examples{
s_s(x = attenu$mag)
s_s(x = attenu)
}
\seealso{
\code{\link{s_mode} \link{s_unique}}
}
