% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_s.R
\name{s_s}
\alias{s_s}
\title{Summary statistics.}
\usage{
s_s(x)
}
\arguments{
\item{x}{a data-frame or vector or factor.}
}
\value{
A base-R data-frame with the following columns:

var: variable names, included only if \code{x} is a data-frame.

n_miss: numbers of missing data.

pct_miss: missing data as percents rounded to the nearest integers.

n_unique: numbers of unique values.

pct_unique: unique values as percents rounded to the nearest integers.

min, max, median, mean: if the data is non-numeric, \code{\link{NA}} is returned.

mode: if \code{\link{NA}} is the most frequent value in the data, \code{\link{NA}} is returned.
If the data has multiple modes, "> 1 mode" is returned.
If the data has no mode, "no mode" is returned.
}
\description{
Tabulate summary statistics of data-frames, vectors, and factors.
}
\examples{
s_s(x = attenu$mag)
s_s(x = attenu)
}
\seealso{
\code{\link{s_cp} \link{s_mode}}
}
