% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_s.R
\name{s_s}
\alias{s_s}
\title{Summary statistics.}
\usage{
s_s(x)
}
\arguments{
\item{x}{a data-frame, vector, factor, or one or more dates / date-times.}
}
\value{
A base-R data-frame with the following columns:

var: variable names, included only if \code{x} is a data-frame.

n_miss: numbers of missing data.

pct_miss: missing data as percents rounded to one decimal place.

n_unique: numbers of unique values.

pct_unique: unique values as percents rounded to one decimal place.

min, max, median, mean: if a variable is non-numeric, \code{\link{NA}} is returned.

mode: if \code{\link{NA}} is the most frequent element in a variable, \code{\link{NA}} or "<NA>" is returned.
If a variable has multiple modes, "s_mode()" is returned.
If a variable has one element, "no mode" is returned.
If a variable has multiple elements and no mode, "no mode" is returned.
}
\description{
Tabulate summary statistics of data-frames, vectors, factors, and dates / date-times.
}
\examples{
zhaoy::s_s(x = attenu$mag)
zhaoy::s_s(x = attenu)
}
\seealso{
\code{\link{s_mode} \link{s_unique}}
}
