% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_s.R
\name{s_s}
\alias{s_s}
\title{Summary Statistics}
\usage{
s_s(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
A tibble.

If \code{x}, or columns in \code{x}, are missing-data, the min / max / median / mode / mean are \code{NA}s.

min and max: If \code{x}, or columns in \code{x}, are not numeric or date / date-time objects, the results are \code{NA}s.

median and mean: If \code{x}, or columns in \code{x}, are not numeric, the results are \code{NA}s.

mode:

If the mode is missing-data, the result is \code{NA}.

If multiple modes exist, the result is "s_mode()".
}
\description{
Tabulate summary statistics.
}
\examples{
zhaoy::s_s(x = attenu$mag)
zhaoy::s_s(x = attenu)
}
\seealso{
\code{\link{s_mode} \link{s_unique}}
}
