% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_s.R
\name{s_s}
\alias{s_s}
\title{Summary statistics of data-frames and vectors.}
\usage{
s_s(x)
}
\arguments{
\item{x}{a data-frame or vector.}
}
\value{
A data-frame with the following columns:

col: column names that display only when the input is a data-frame.

n_miss: numbers of missing data.

pct_miss: percents of data that are missing data, rounded to the nearest integers.

n_unique: numbers of unique values.

pct_unique: percents of data that are unique values, rounded to the nearest integers.

min, max, median, mean: \code{\link{NA}} for non-numeric data.

mode: \code{\link{NA}} only when \code{\link{NA}} is the most frequent value.
When the data has multiple modes, the return value is "> 1 mode".
When the data has no mode, the return value is "no mode".
}
\description{
Tabulate summary statistics of data-frames and vectors.
}
\examples{
s_s(x = attenu$station)
s_s(x = attenu)
}
\seealso{
\code{\link{s_cp} \link{s_mode}}
}
