% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_tbl.R
\name{s_tbl}
\alias{s_tbl}
\title{Column or table summaries}
\usage{
s_tbl(x)
}
\arguments{
\item{x}{a column or table.}
}
\value{
A table.
}
\description{
Tabulates summary statistics of columns or tables.
}
\details{
The outputs are:

col: column name, displayed only when input is a table

n_miss: number of missing data

pct_miss: percent of a column that are missing data

n_unique: number of unique values in a column

pct_unique: percent of a column that are unique values

min, max, median, mean: \code{\link{NA}} when a column is non-numeric

mode: NA only when NA is the most frequent value in a column
}
\examples{
s_tbl(x = attenu$station)
s_tbl(x = attenu)
}
\seealso{
\code{\link{s_col} \link{s_mode}}
}
