% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_tbl.R
\name{s_tbl}
\alias{s_tbl}
\title{Summarize vectors or data-frames}
\usage{
\code{s_tbl(x)}
}
\arguments{
\item{x}{a vector or data-frame.}
}
\value{
A data-frame.
}
\description{
Tabulate summary statistics of vectors or data-frames, which are made up of vectors.
}
\details{
The contents of the output are:

col: column name, displayed only if the input is a data-frame

n_miss: number of missing values

pct_miss: percent of column that are missing values

n_unique: number of unique values

pct_unique: percent of column that are unique values

min, max, median, mean
}
\examples{
s_tbl(x = airquality)
s_tbl(x = airquality$Ozone)
}
\seealso{
\code{\link{s_col}}
}
