% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_tbl.R
\name{s_tbl}
\alias{s_tbl}
\title{Column or table summaries}
\usage{
s_tbl(x)
}
\arguments{
\item{x}{a column or table for which a summary is desired.}
}
\value{
A table.
}
\description{
Tabulate summary statistics of columns or tables.
}
\details{
The outputs are:

col: column name, displayed only if the input is a table

n_miss: number of missing data

pct_miss: percent of column that are missing data

n_unique: number of unique values

pct_unique: percent of column that are unique values

min, max, median, mode, mean
}
\examples{
s_tbl(x = attenu$station)
s_tbl(x = attenu)
}
\seealso{
\code{\link{s_col}}
}
