% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_tbl.R
\name{s_tbl}
\alias{s_tbl}
\title{Column or table summaries}
\usage{
s_tbl(x)
}
\arguments{
\item{x}{a column or table for which a summary is desired.}
}
\value{
A table.
}
\description{
Tabulates summary statistics of columns or tables.
}
\details{
The outputs are:

col: column name, displayed only when input is a table

n_miss: number of missing data

pct_miss: missing data as percent of column

n_unique: number of unique values

pct_unique: unique values as percent of column

min, max, median, mean: \code{\link{NA}} when column is non-numeric

mode: NA only when NA is the most frequent value in column
}
\examples{
s_tbl(x = attenu$station)
s_tbl(x = attenu)
}
\seealso{
\code{\link{s_col} \link{s_mode}}
}
