% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_unique.R
\name{s_unique}
\alias{s_unique}
\title{Summarize unique elements.}
\usage{
s_unique(x)
}
\arguments{
\item{x}{a vector, factor, or one or more dates / date-time(s).}
}
\value{
A base-R data-frame with the following columns:

element: unique elements,
beginning with "<NA>" if \code{\link{NA}} is present,
then continuing in ascending order.

n: counts.

pct: counts as percents rounded to one decimal place.
}
\description{
Tabulate counts and percents of unique elements within vectors, factors, and dates / date-time(s).
}
\examples{
s_unique(x = attenu$station)
}
\seealso{
\code{\link{s_mode} \link{s_s}}
}
