% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_unique.R
\name{s_unique}
\alias{s_unique}
\title{Unique Values}
\usage{
s_unique(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A tibble.
}
\description{
Tabulate counts and percents of unique values, including missing-data.
}
\examples{
s_unique(x = attenu$station)
}
\seealso{
\code{\link{s_mode} \link{s_s}}
}
