% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_unique.R
\name{s_unique}
\alias{s_unique}
\title{Frequency of unique values}
\usage{
s_unique(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A data-frame.
}
\description{
Tabulate counts and percents of unique values, including missing-data, in vectors.
}
\examples{
zhaoy::s_unique(x = attenu$station)
}
