% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_vec.R
\name{s_vec}
\alias{s_vec}
\title{Summarize vectors or data-frames.}
\usage{
\code{s_vec(x)}
}
\arguments{
\item{x}{a data-frame.}
}
\value{
A data-frame.
}
\description{
Tabulate summary statistics of vectors or data-frames, which are made up of vectors.
}
\details{
The contents of the output are:

col: column name

n_miss: number of missing values

pct_miss: missing values as percent of column

n_unique: number of unique values

pct_unique: unique values as percent of column

min, max, median, mean
}
\examples{
s_vec(x = airquality)
}
\seealso{
\code{\link{s_value}}
}
