% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lab_pe.R
\name{tidy_lab_pe}
\alias{tidy_lab_pe}
\title{Tidy Provide Enterprise laboratory data.}
\usage{
tidy_lab_pe(x)
}
\arguments{
\item{x}{A data-frame.}
}
\value{
A tibble.
}
\description{
Tidy the Provide Enterprise "Test Results by Client With ID" report.
}
