% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pel.R
\name{tidy_pel}
\alias{tidy_pel}
\title{Tidy Provide Enterprise laboratory data.}
\usage{
tidy_pel(folder, path, sheet = NULL, range = NULL)
}
\arguments{
\item{folder}{any folder above both 1) the .xlsx file and 2) the R file.}

\item{path}{relative to \code{folder}, path to the .xlsx file.}

\item{sheet}{sheet to read.}

\item{range}{a cell range to read from.}
}
\value{
A data-frame.
}
\description{
Tidy the .xlsx version of the Provide Enterprise "Test Results by Client with ID" report.
}
