% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pel.R
\name{tidy_pel}
\alias{tidy_pel}
\title{Convert PE lab data into tidy data frames}
\usage{
tidy_pel(folder, path, sheet = NULL)
}
\arguments{
\item{folder}{Folder above the xls / xlsx file.}

\item{path}{Path to the file, excluding the folder.}

\item{sheet}{Sheet to read.}
}
\value{
A table.
}
\description{
Tidies PE lab data that are in Microsoft Excel files.
}
