% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pel.R
\name{tidy_pel}
\alias{tidy_pel}
\title{Convert PE lab data into tidy data frames}
\usage{
tidy_pel(criterion, file, sheet)
}
\arguments{
\item{criterion}{A criterion.}

\item{file}{Name and extension of the xlsx file.}

\item{sheet}{Sheet to read.
Either a string (the name of a sheet), or an integer (the position of the sheet).}
}
\description{
Tidies PE lab data that were exported in xlsx files.
}
