% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_lc_df.R
\name{zhaoy_lc_df}
\alias{zhaoy_lc_df}
\title{Character case}
\usage{
zhaoy_lc_df(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A vector of the same length as \code{x}.
}
\description{
Translate characters from upper to lower case. Leave non-alphabetic characters un-changed.
}
