% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_lz_id.R
\name{zhaoy_lz_id}
\alias{zhaoy_lz_id}
\title{Leading zeros in numeric identifiers.}
\usage{
zhaoy_lz_id(x, lz)
}
\arguments{
\item{x}{a vector of length one, up to 9 digits.}

\item{lz}{Logical: include leading zeros?}
}
\value{
A character vector of length one.
}
\description{
Includes or excludes leading zeros in numeric identifiers.
}
\examples{
zhaoy_lz_id(x = 1, lz = TRUE)
zhaoy_lz_id(x = "000000001", lz = FALSE)
}
