% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_lz_id.R
\name{zhaoy_lz_id}
\alias{zhaoy_lz_id}
\title{Leading zeros in integer identifiers.}
\usage{
zhaoy_lz_id(x, lz)
}
\arguments{
\item{x}{a character, factor, or numeric vector of one up-to-9-digits integer.}

\item{lz}{logical: TRUE includes, and FALSE excludes, leading zeros.}
}
\value{
A character vector of length one.
}
\description{
Include or exclude leading zeros in integer identifiers.
}
\examples{
zhaoy_lz_id(x = 1, lz = TRUE)
zhaoy_lz_id(x = "000000001", lz = FALSE)
}
