% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_lz_id.R
\name{zhaoy_lz_id}
\alias{zhaoy_lz_id}
\title{Leading zeros in identifiers.}
\usage{
zhaoy_lz_id(x, lz)
}
\arguments{
\item{x}{a vector or factor of one identifier.}

\item{lz}{logical: TRUE includes, and FALSE excludes, leading zeros.}
}
\value{
A length-one character vector.
}
\description{
Includes or excludes leading zeros in identifiers that have 1 - 9 digits,
of which at least one digit is a positive integer
and any remaining digits are either zeros or positive integers.
}
\examples{
zhaoy_lz_id(x = 1, lz = TRUE)
zhaoy_lz_id(x = "000000001", lz = FALSE)
}
