% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_lz_id.R
\name{zhaoy_lz_id}
\alias{zhaoy_lz_id}
\title{Leading Zeros in Epic ID Numbers}
\usage{
zhaoy_lz_id(x, lz)
}
\arguments{
\item{x}{an Epic ID number.}

\item{lz}{logical: TRUE includes, and FALSE excludes, leading zeros.}
}
\value{
A length-one character vector.
}
\description{
Include or exclude leading zeros in Epic ID numbers.

Epic ID numbers have 1-8 positive integers and 0-8 leading zeros.
}
