% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_lz_id.R
\name{zhaoy_lz_id}
\alias{zhaoy_lz_id}
\title{Leading zeros in identifiers.}
\usage{
zhaoy_lz_id(x, lz)
}
\arguments{
\item{x}{a vector or factor of one numbers-only identifier that has 1 - 9 digits.}

\item{lz}{logical: TRUE includes, and FALSE excludes, leading zeros.}
}
\value{
A length-one character vector.
}
\description{
Includes or excludes leading zeros in number-only identifiers of 1 - 9 digits.
}
\examples{
zhaoy_lz_id(x = 1, lz = TRUE)
zhaoy_lz_id(x = "000000001", lz = FALSE)
}
