% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_mode.R
\name{zhaoy_mode}
\alias{zhaoy_mode}
\title{Statistical mode.}
\usage{
zhaoy_mode(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A vector of length one.
If there are multiple modes, the return value is "> 1 mode".
If there is no mode, the return value is "no mode".
}
\description{
Calculate the statistical mode.
}
\examples{
zhaoy_mode(x = mtcars$mpg)
}
