% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_mode.R
\name{zhaoy_mode}
\alias{zhaoy_mode}
\title{Statistical mode.}
\usage{
zhaoy_mode(x)
}
\arguments{
\item{x}{a vector, factor, or one or more dates / date-times.}
}
\value{
A length-one vector.

If \code{\link{NA}} is the most frequent element in \code{x}, \code{\link{NA}} is returned.

If \code{x} has multiple modes, "s_mode()" is returned.

If \code{x} has one element, "no mode" is returned.

If \code{x} has multiple elements and no mode, "no mode" is returned.
}
\description{
Calculate the statistical mode.
}
\examples{
zhaoy_mode(x = attenu$dist)
}
