% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_mode.R
\name{zhaoy_mode}
\alias{zhaoy_mode}
\title{Statistical mode.}
\usage{
zhaoy_mode(x)
}
\arguments{
\item{x}{a vector or factor.}
}
\value{
A vector of length one.

If \code{\link{NA}} is the most frequent value in \code{x}, "NA" or "<NA>" is returned.

If \code{x} has multiple modes, "> 1 mode" is returned.

If \code{x} has no mode, "no mode" is returned.
}
\description{
Calculate the statistical mode.
}
\examples{
zhaoy_mode(x = attenu$dist)
}
