% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_mode.R
\name{zhaoy_mode}
\alias{zhaoy_mode}
\title{Statistical mode.}
\usage{
zhaoy_mode(x)
}
\arguments{
\item{x}{a vector, factor, or one or more dates / date-times.}
}
\value{
A vector of length one.

If \code{\link{NA}} is the most frequent value in \code{x}, \code{\link{NA}} is returned.

If \code{x} has multiple modes, "> 1 mode" is returned.

If \code{x} has a length of one, or has otherwise no mode, "no mode" is returned.
}
\description{
Calculate the statistical mode.
}
\examples{
zhaoy_mode(x = attenu$dist)
}
