% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_numeric.R
\name{zhaoy_numeric}
\alias{zhaoy_numeric}
\title{Summary statistics of numeric data.}
\usage{
zhaoy_numeric(x, fun)
}
\arguments{
\item{x}{a vector.}
}
\value{
A vector.
}
\description{
If data are numeric, calculate the minimum, maximum, median, and mean.
}
