% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_mode.R
\name{zhaoy_s_mode}
\alias{zhaoy_s_mode}
\title{Statistical mode}
\usage{
zhaoy_s_mode(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A length-one vector.

If the mode is missing-data: \code{NA}.

If multiple modes exist: "s_mode()".
}
\description{
Calculate the statistical mode.
}
