% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_mode.R
\name{zhaoy_s_mode}
\alias{zhaoy_s_mode}
\title{Statistical mode.}
\usage{
zhaoy_s_mode(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
A length-one character vector.

If the mode is \code{\link{NA}}, \code{\link{NA}} is returned.

If multiple modes exist, "s_mode()" is returned.
}
\description{
Compute the statistical mode.
}
