% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_mode.R
\name{zhaoy_s_mode}
\alias{zhaoy_s_mode}
\title{Statistical Mode}
\usage{
zhaoy_s_mode(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
A length-one character vector.

If no mode exists, the result is "no mode".

If the mode is missing-data, the result is \code{NA}.

If multiple modes exist, the result is "s_mode()".
}
\description{
Compute the statistical mode.
}
