% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_numeric.R, R/zhaoy_s_numeric.R
\name{zhaoy_s_numeric}
\alias{zhaoy_s_numeric}
\alias{zhaoy_s_numeric}
\title{Summary statistics of numeric data.}
\usage{
zhaoy_s_numeric(x, fun)

zhaoy_s_numeric(x, fun)
}
\arguments{
\item{x}{a numeric vector.}

\item{x}{a numeric vector.}
}
\value{
A vector of length one.

A vector of length one.
}
\description{
If data are numeric, calculate the minimum, maximum, median, and mean.

If data are numeric, calculate the minimum, maximum, median, or mean.
}
\examples{
zhaoy_s_numeric(x = mtcars$mpg, fun = "min")
zhaoy_s_numeric(x = mtcars$mpg, fun = "min")
}
