% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_numeric.R
\name{zhaoy_s_numeric}
\alias{zhaoy_s_numeric}
\title{Summary statistics.}
\usage{
zhaoy_s_numeric(x, s)
}
\arguments{
\item{x}{a vector, factor, or one or more dates / date-times.}
}
\value{
A vector.

If \code{x} is numeric, a number is returned.

If \code{x} is non-numeric, \code{\link{NA}} is returned.
}
\description{
For numeric data, calculate the minimum, maximum, median, or mean.
}
\examples{
zhaoy_s_numeric(x = attenu$accel, s = "min")
}
