% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_numeric.R
\name{zhaoy_s_numeric}
\alias{zhaoy_s_numeric}
\title{Summary statistics of numeric data.}
\usage{
zhaoy_s_numeric(x, fun)
}
\arguments{
\item{x}{a vector.}
}
\value{
If \code{x} is numeric, a vector of length one.

If \code{x} is non-numeric, a vector of one \code{\link{NA}}.
}
\description{
For numeric data, calculate the minimum, maximum, median, or mean.
}
\examples{
zhaoy_s_numeric(x = attenu$accel, fun = "min")
}
