% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_s.R
\name{zhaoy_s_s}
\alias{zhaoy_s_s}
\title{Summary Statistics}
\usage{
zhaoy_s_s(x, s)
}
\arguments{
\item{x}{an R object.}

\item{s}{a statistic.}
}
\value{
A length-one vector.

The result is \code{NA} if \code{x} is:

missing-data

a date / date-time / time-interval object,
and \code{s = "median"} or \code{s = "mean"}

not a numeric or date / date-time / time-interval object.
}
\description{
Compute the minimum, maximum, median, or mean.
}
