% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_s_s.R
\name{zhaoy_s_s}
\alias{zhaoy_s_s}
\title{Summary Statistics}
\usage{
zhaoy_s_s(x, s)
}
\arguments{
\item{x}{an R object.}

\item{s}{a statistic.}
}
\value{
A length-one character vector.

If \code{x} is not a numeric or date / date-time / time-interval object,
and \code{x = "min"} or \code{x = "max"},
\code{\link{NA}} is returned.

If \code{x} is a date / date-time / time-interval object,
and \code{x = "median"} or \code{x = "mean"},
\code{\link{NA}} is returned.
}
\description{
Compute the minimum, maximum, median, or mean.
}
