% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_tolower.R
\name{zhaoy_tolower}
\alias{zhaoy_tolower}
\title{Convert to lower-case.}
\usage{
zhaoy_tolower(x)
}
\arguments{
\item{x}{a vector, factor, or one or more Date / POSIXlt / POSIXct objects.}
}
\value{
If \code{x} is an alphabetic vector or factor, a character vector of the same length as \code{x}.

Otherwise, \code{x} un-changed.
}
\description{
Convert upper-case characters and factor levels to lower-case characters.

Return non-alphabetic vectors and factors, and Date / POSIXlt / POSIXct objects, un-changed.
}
\examples{
x <- c("MiXeD", "cAsE", "123")
zhaoy_tolower(x = x)
}
