% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_tolower.R
\name{zhaoy_tolower}
\alias{zhaoy_tolower}
\title{Convert to lower-case.}
\usage{
zhaoy_tolower(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A vector of the same length as \code{x}.
}
\description{
Convert upper-case letters to lower-case letters.

Return non-alphabetical objects un-changed.
}
