% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_tolower.R
\name{zhaoy_tolower}
\alias{zhaoy_tolower}
\title{Convert characters and factors to lower-case.}
\usage{
zhaoy_tolower(x)
}
\arguments{
\item{x}{a character or factor vector.}
}
\value{
A character or factor vector of the same length as \code{x}.
}
\description{
Execute \code{\link{tolower}} on character and factor vectors.
Other vectors are left un-changed.
}
\examples{
x <- c("MiXeD", "cAsE", "123")
zhaoy_tolower(x = x)
}
