% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhaoy_tolower.R
\name{zhaoy_tolower}
\alias{zhaoy_tolower}
\title{Convert characters and factors to lower-case.}
\usage{
zhaoy_tolower(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
a vector.
}
\description{
If data are characters or factors, execute \code{\link{tolower}}.
}
